Shrink data size by removing repeated data and removed view model objects from the saved instance state and arguments to reduce size.

+++ app/src/main/java/delit/piwigoclient/piwigoApi/upload/BasePiwigoUploadService.java
@@ -100,7 +100,7 @@ public abstract class BasePiwigoUploadService extends JobIntentService {
     }
 
     public static @NotNull
-    UploadJob createUploadJob(ConnectionPreferences.ProfilePreferences connectionPrefs, ArrayList<File> filesForUpload, CategoryItemStub category, boolean compressVideosBeforeUpload, int uploadedFilePrivacyLevel, long responseHandlerId) {
+    UploadJob createUploadJob(ConnectionPreferences.ProfilePreferences connectionPrefs, ArrayList<File> filesForUpload, CategoryItemStub category, boolean compressVideosBeforeUpload, byte uploadedFilePrivacyLevel, long responseHandlerId) {
         long jobId = getNextMessageId();
         UploadJob uploadJob = new UploadJob(connectionPrefs, jobId, responseHandlerId, filesForUpload, category, compressVideosBeforeUpload, uploadedFilePrivacyLevel);
         synchronized (activeUploadJobs) {
